requirements.txt
fastapi 
uvicorn 
openai 
python-dotenv 
reportlab 
pydantic 
aiofiles
groq
azure-communication-email